Business logic change: users may now edit/delete busy datasets Fix tests, tidy, integrate with UI
